Introducing the New Tweak.
Very much like the old tweak but with some important changes.

All modifiers that activate the tweak tools are now user defined. You
will need to assign bind keys to the tweak tools, and hotkeys to the
magnet options and constraint toggles. I've included help messages in
this release. I will probably add more help as questions arise. The help
can be turned off.

The old tweak is still available in the usual place. The new tweak is
activated through its own menu. The new tweak menu is listed in the
Info Line and is shown by pressing Alt+Rmb (Alt+Ctrl for Maya cam users).
Be sure to read the help strings on how to bind keys to the various tweak
tools. Briefly... while hovering over the cmd in the tweak menu, press
and hold the desired mod keys (ctrl,shift,alt) in combination with the
mouse button that will activate the cmd. There is an option to swaop keys
and another to delete bind keys(Rmb).

The Magnet radius is now adjusted by holding Alt and moving the mouse left
and right. F1-3 still activate temporary constraints, however toggles have
to be hotkeyed in Rmb|TweakMenu|XYZ Constraints.