Add we() to wings_draw_setup
This makes it possibly for plugins to get/reuse the needed drawing
information.

I haven't really decided on what it should return yet, we might want
to use vbos later, which makes less copies of the data in the drivers.
In that case the user shouldn't use the buffers directly.